+++ R/exact_extract_helpers.R
@@ -180,6 +180,36 @@
   return(r)
 }
 
+.crs <- function(r) {
+  if(inherits(r, 'BasicRaster')) {
+    if (utils::packageVersion('raster') < numeric_version('3.5')) {
+      return(raster::crs(r))
+    } else {
+      return(terra::crs(r))
+    }
+  } else if (inherits(r, 'SpatRaster')) {
+    return(terra::crs(r))
+  } else {
+    stop('Unknown type: ', class(r))
+  }
+}
+
+.setValues <- function(r, x) {
+  if(inherits(r, 'BasicRaster')) {
+    if (utils::packageVersion('raster') < numeric_version('3.5')) {
+      raster::values(r) <- x
+    } else {
+      terra::values(r) <- x
+    }
+  } else if (inherits(r, 'SpatRaster')) {
+    raster::values(r) <- x
+  } else {
+    stop('Unknown type: ', class(r))
+  }
+
+  return(r)
+}
+
 .numLayers <- function(r) {
   if(inherits(r, 'BasicRaster')) {
     return(raster::nlayers(r))
@@ -204,11 +234,31 @@
   }
 }
 
+.colFromX <- function(r, x) {
+  if (inherits(r, 'BasicRaster')) {
+    raster::colFromX(r, x)
+  } else if (inherits(r, 'SpatRaster')) {
+    terra::colFromX(r, x)
+  } else {
+    stop('Unknown type: ', class(r))
+  }
+}
+
 .yFromRow <- function(r, row) {
   if (inherits(r, 'BasicRaster')) {
     raster::yFromRow(r, row)
   } else if (inherits(r, 'SpatRaster')) {
     terra::yFromRow(r, row)
+  } else {
+    stop('Unknown type: ', class(r))
+  }
+}
+
+.rowFromY <- function(r, y) {
+  if (inherits(r, 'BasicRaster')) {
+    raster::rowFromY(r, y)
+  } else if (inherits(r, 'SpatRaster')) {
+    terra::rowFromY(r, y)
   } else {
     stop('Unknown type: ', class(r))
   }